chore : rename directories to plural
